requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-chroma>=0.2
langchain-community>=0.3
langchain-huggingface>=0.1
langgraph>=0.3
langfuse>=3.0
chromadb>=0.6
sentence-transformers>=3.0
pydantic>=2.0
pydantic-settings>=2.0
pymupdf>=1.24
pdfplumber>=0.11
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
openai>=1.0
ragas>=0.2
requests>=2.32
